% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine.R
\name{digest_simple}
\alias{digest_simple}
\title{Simple string hash for font dedup}
\usage{
digest_simple(x)
}
\arguments{
\item{x}{Character string.}
}
\value{
Character hash.
}
\description{
Simple string hash for font dedup
}
\keyword{internal}
