% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pikchrWrappers.R
\docType{data}
\name{google_fonts}
\alias{google_fonts}
\title{Google Font List}
\format{
## `google_fonts`
A data frame with 1,718 rows and 2 columns:
\describe{
  \item{family}{Font family names}
  \item{styles}{Font styles}
}
}
\source{
<https://fonts.google.com>
}
\usage{
google_fonts
}
\description{
List of fonts and its stytles on google fonts site.
Report ...
}
\keyword{datasets}
