% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine.R
\name{run_hook_plot}
\alias{run_hook_plot}
\title{Run Plot Hook and Record Plot Filenames}
\usage{
run_hook_plot(x, options)
}
\arguments{
\item{x}{A character string representing the path to the plot file.}

\item{options}{A list of options associated with the Knitr chunk, such as plot settings and output options.}
}
\value{
The result of calling the current Knitr plot hook with the specified plot file and options.
}
\description{
This function records the filenames of plot outputs (such as SVG or PNG files) in the Knitr environment 
and then runs the appropriate plot hook to display or save the plots.
}
\details{
The `run_hook_plot` function appends the plot file path to the internal Knitr option `plot_files`. 
It then retrieves the current plot hook defined in `knit_hooks` and calls it to handle the plot rendering or output. 
This function is particularly useful for handling plots generated from both R code and other outputs like HTML widgets.
}
\keyword{internal}
